# internal/web/api/api.yml
openapi: 3.1.1
info:
  title: Mahiron API
  description: Yet another DVR Tuner Server for Japanese TV
  contact:
    name: 21S1298001
    url: https://github.com/21S1298001
  version: 5.0.0
servers:
  - url: http://localhost:40772/api
...
paths:
...
  /events/stream:
    parameters: []
    get:
      tags:
        - events
        - stream
      operationId: getEventsStream
      parameters:
        - name: resource
          in: query
          schema:
            type: string
            enum:
              - program
              - service
              - tuner
              - job
              - job_schedule
        - name: type
          in: query
          schema:
            type: string
            enum:
              - create
              - update
              - remove
      responses:
        "200":
          description: OK
          content:
            text/plain:
              x-ogen-raw-response: true
              schema:
                type: string
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /events:
    parameters: []
    get:
      tags:
        - events
      operationId: getEvents
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
...
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: integer
        reason:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorOfOpenAPI"
    ErrorOfOpenAPI:
      type: object
      properties:
        errorCode:
          type: string
        message:
          type: string
        location:
          type: string
...
    UnixtimeMS:
      type: integer
...
    Event:
      required:
        - data
        - resource
        - time
        - type
      type: object
      properties:
        resource:
          $ref: "#/components/schemas/EventResource"
        type:
          $ref: "#/components/schemas/EventType"
        data:
          type: object
          additionalProperties: {}
        time:
          $ref: "#/components/schemas/UnixtimeMS"
    EventResource:
      type: string
      enum:
        - program
        - service
        - tuner
        - job
        - job_schedule
    EventType:
      type: string
      enum:
        - create
        - update
        - remove
...